requirements.txt
google-api-python-client
pandas
praw
numpy
nltk
ntscraper
matplotlib
seaborn
python-dotenv
-e .